requirements.txt
absl-py==1.4.0
aiobotocore==2.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
asn1crypto==1.5.1
atpublic==5.0
attrs==24.3.0
blinker==1.9.0
botocore==1.35.93
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
cryptography==44.0.0
dataclasses-json==0.6.7
eval_type_backport==0.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2023.12.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langsmith==0.2.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
mistralai==1.3.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.22.0
numpy==1.26.4
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.0
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
s3fs==2023.12.2
scikit-learn==1.5.2
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake==1.0.2
snowflake-connector-python==3.12.4
snowflake-ml-python==1.7.1
snowflake-snowpark-python==1.26.0
snowflake._legacy==1.0.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
streamlit==1.41.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
watchdog==6.0.0
wrapt==1.17.2
xgboost==2.1.3
yarl==1.18.3
